WIP Xorg user session
We want to support Xorg user session and keep running it as root as an
option, so we don't break the workflow of some users.

As a matter of fact, when we run the X11 server as sddm user, we also
start the display start and stop scripts as user.  Scripts that need
root privileges won't work.

To support both modes, we move the code to sddm-x11-helper and either
run it as root or sddm user using sddm-helper.

Closes: #246